pyproject.toml
[tool.poetry]
name = "licenses-everywhere"
version = "0.1.0"
description = "A tool to ensure all public repositories in a GitHub organization have proper license files"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/yourusername/licenses-everywhere"
packages = [{include = "licenses_everywhere"}]

[tool.poetry.dependencies]
python = "^3.8.1"
PyGithub = "^1.58.2"
click = "^8.1.3"
rich = "^13.3.5"
Jinja2 = "^3.1.2"
pyyaml = "^6.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
licenses-everywhere = "licenses_everywhere.cli:main"

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true 